fix: debug playwright workflow
