+++ tests/frontend.spec.ts
@@ -1,42 +1,5 @@
 import { test, expect } from "@playwright/test";
-
-async function openRegisterModal(page: import("@playwright/test").Page) {
-  await page.goto("/");
-  await page.getByRole("button", { name: /log in/i }).click();
-  await page.getByRole("button", { name: /register here/i }).click();
-  await expect(page.getByRole("heading", { name: /register/i })).toBeVisible();
-}
-
-async function openLoginModal(page: import("@playwright/test").Page) {
-  await page.goto("/");
-  await page.getByRole("button", { name: /log in/i }).click();
-  await expect(page.getByRole("heading", { name: /login/i })).toBeVisible();
-}
-
-async function registerAndLogin(page: import("@playwright/test").Page) {
-  const ts = Date.now();
-  const credentials = {
-    email: `fe-user-${ts}@example.com`,
-    username: `feuser${ts}`,
-    password: "SecurePass123!",
-  };
-
-  // Register
-  await openRegisterModal(page);
-  await page.fill("#email", credentials.email);
-  await page.fill("#username", credentials.username);
-  await page.fill("#password", credentials.password);
-  await page.fill("#confirmPassword", credentials.password);
-  await page.getByRole("button", { name: /^submit$/i }).click();
-
-  // Login
-  await openLoginModal(page);
-  await page.fill("#email", credentials.email);
-  await page.fill("#password", credentials.password);
-  await page.getByRole("button", { name: /^submit$/i }).click();
-
-  await expect(page.getByText(credentials.username)).toBeVisible();
-}
+import { registerAndLogin } from "./utils";
 
 test("home page loads", async ({ page }) => {
   await page.goto("/");
